+++ mirrord-layer/src/pod_api.rs
@@ -252,8 +252,8 @@ impl KubernetesAPI {
 
     async fn create_job_pod_agent(
         &self,
-        agent_image: String,
         runtime_data: RuntimeData,
+        agent_image: String,
         connection_port: u16,
         progress: &TaskProgress,
     ) -> Result<String> {
@@ -394,7 +394,7 @@ impl KubernetesAPI {
             )
             .await?
         } else {
-            self.create_job_pod_agent(agent_image, runtime_data, connection_port, &progress)
+            self.create_job_pod_agent(runtime_data, agent_image, connection_port, &progress)
                 .await?
         };
         progress.done_with("agent running");
@@ -691,7 +691,9 @@ impl FromStr for Target {
     fn from_str(target: &str) -> Result<Target> {
         let mut split = target.split('/');
         match split.next() {
-            Some("deployment") => DeploymentTarget::from_split(&mut split).map(Target::Deployment),
+            Some("deployment") | Some("deploy") => {
+                DeploymentTarget::from_split(&mut split).map(Target::Deployment)
+            }
             Some("pod") => PodTarget::from_split(&mut split).map(Target::Pod),
             _ => Err(LayerError::InvalidTarget(format!(
                 "Provided target: {:?} is neither a pod or a deployment.",